+++ core/src/main/java/com/agifans/agile/GameState.java
@@ -80,7 +80,7 @@ public class GameState {
      * are the ARGB values. The dimensions of this are 320x168, i.e. two pixels per 
      * AGI pixel. Makes it easier to copy to the main pixels array when required.
      */
-    public short[] visualPixels;
+    public int[] visualPixels;
 
     /**
      * The pixel array for the priority data for the current Picture, where the values
@@ -188,7 +188,7 @@ public class GameState {
         this.recognisedWords = new ArrayList<>();
         this.scriptBuffer = new ScriptBuffer(this);
 
-        this.visualPixels = new short[320 * 168];
+        this.visualPixels = new int[320 * 168];
         this.priorityPixels = new int[160 * 168];
         this.controlPixels = new int[160 * 168];
 
@@ -334,7 +334,7 @@ public class GameState {
      * 
      * @param pixels The screen pixels to blit the AnimatedObjects to.
      */
-    public void showObjects(short[] pixels) {
+    public void showObjects(int[] pixels) {
         // If no list specified, then draw stopped list then update list.
         showObjects(pixels, stoppedObjectList);
         showObjects(pixels, updateObjectList);
@@ -348,7 +348,7 @@ public class GameState {
      * @param pixels The screen pixels to blit the AnimatedObjects to.
      * @param objectShowList
      */
-    public void showObjects(short[] pixels, List<AnimatedObject> objectShowList) {
+    public void showObjects(int[] pixels, List<AnimatedObject> objectShowList) {
         for (AnimatedObject aniObj : objectShowList)
         {
             aniObj.show(pixels);
